CUDA changes for rc3.10

+++ nd4j-backends/nd4j-backend-impls/nd4j-cuda-7.5/src/main/java/org/nd4j/linalg/jcublas/buffer/CudaFloatDataBuffer.java
@@ -105,7 +105,7 @@ public class CudaFloatDataBuffer extends BaseCudaDataBuffer {
     }
 
     public CudaFloatDataBuffer(ByteBuffer buffer, long length) {
-        super(buffer, length);
+        super(buffer, (int) length);
     }
 
     public CudaFloatDataBuffer(ByteBuffer buffer, long length, int offset) {
